corrección conceptos de pago

+++ WebMigradorCtasPorCobrar/WebMigradorCtasPorCobrar/Models/Entities/TemporalPagos/ConceptoPago.cs
@@ -7,7 +7,7 @@ namespace WebMigradorCtasPorCobrar.Models.Entities.TemporalPagos
 {
     public class ConceptoPago
     {
-        public double Id_cp { get; set; }
+        public int Id_cp { get; set; }
         public int Cuota_pago { get; set; }
         public string Ano { get; set; }
         public string P { get; set; }
@@ -18,8 +18,8 @@ namespace WebMigradorCtasPorCobrar.Models.Entities.TemporalPagos
         public string Clasific_5 { get; set; }
         public double Id_cp_agrp { get; set; }
         public bool Agrupa { get; set; }
-        public double Nro_pagos { get; set; }
-        public double Id_cp_afec { get; set; }
+        public int Nro_pagos { get; set; }
+        public int Id_cp_afec { get; set; }
         public bool Porcentaje { get; set; }
         public decimal Monto { get; set; }
         public bool Eliminado { get; set; }
